+++ blocklets/ai-studio/src/components/file-editor/execute-block.tsx
@@ -30,6 +30,7 @@ import Add from '../../pages/project/icons/add';
 import External from '../../pages/project/icons/external';
 import Trash from '../../pages/project/icons/trash';
 import { PROMPTS_FOLDER_NAME, useCreateFile, useProjectStore } from '../../pages/project/yjs-state';
+import IndicatorTextField from '../awareness/indicator-text-field';
 import PromptEditorField from './prompt-editor-field';
 
 export default function ExecuteBlockForm({
@@ -38,6 +39,7 @@ export default function ExecuteBlockForm({
   assistant,
   value,
   readOnly,
+  path,
   compareAssistant,
   isRemoteCompare,
   ...props
@@ -46,6 +48,7 @@ export default function ExecuteBlockForm({
   gitRef: string;
   assistant: AssistantYjs;
   value: ExecuteBlockYjs;
+  path: (string | number)[];
   readOnly?: boolean;
   compareAssistant?: AssistantYjs;
   isRemoteCompare?: boolean;
@@ -69,17 +72,30 @@ export default function ExecuteBlockForm({
     <Stack {...props} sx={{ border: 2, borderColor: 'warning.main', borderRadius: 1, p: 1, gap: 1, ...props.sx }}>
       <Stack direction="row" gap={1} alignItems="center">
         <Typography variant="subtitle2">{t('executeBlock')}</Typography>
-
-        <TextField
-          size="small"
-          select
-          hiddenLabel
-          SelectProps={{ autoWidth: true }}
-          value={value.selectType || 'all'}
-          onChange={(e) => (value.selectType = e.target.value as any)}>
-          <MenuItem value="all">{t('all')}</MenuItem>
-          <MenuItem value="selectByPrompt">{t('selectPrompt')}</MenuItem>
-        </TextField>
+        <IndicatorTextField
+          projectId={projectId}
+          gitRef={gitRef}
+          path={[value.id, value.selectType ?? 'all']}
+          TextFiledProps={{
+            size: 'small',
+            select: true,
+            hiddenLabel: true,
+            SelectProps: {
+              autoWidth: true,
+            },
+            value: value.selectType || 'all',
+            onChange: (e: React.ChangeEvent<HTMLInputElement | HTMLTextAreaElement>) =>
+              (value.selectType = e.target.value as any),
+            children: [
+              <MenuItem key="all" value="all">
+                {t('all')}
+              </MenuItem>,
+              <MenuItem key="selectByPrompt" value="selectByPrompt">
+                {t('selectPrompt')}
+              </MenuItem>,
+            ],
+          }}
+        />
 
         <Box flex={1} />
       </Stack>
@@ -87,9 +103,11 @@ export default function ExecuteBlockForm({
       {value.selectType === 'selectByPrompt' && (
         <Stack>
           <Typography variant="caption">{t('prompt')}</Typography>
-
           <PromptEditorField
             readOnly={readOnly}
+            projectId={projectId}
+            gitRef={gitRef}
+            path={path.concat('selectByPrompt')}
             assistant={assistant}
             value={value.selectByPrompt}
             onChange={(prompt) => (value.selectByPrompt = prompt)}
@@ -188,31 +206,50 @@ export default function ExecuteBlockForm({
           {t('formatResult')}
         </Typography>
 
-        <TextField
-          hiddenLabel
-          select
-          SelectProps={{ autoWidth: true }}
-          value={value.formatResultType || 'none'}
-          onChange={(e) => (value.formatResultType = e.target.value as any)}>
-          <MenuItem value="none">{t('stayAsIs')}</MenuItem>
-          <MenuItem value="asHistory">{t('asHistory')}</MenuItem>
-        </TextField>
+        <IndicatorTextField
+          projectId={projectId}
+          gitRef={gitRef}
+          path={[value.id, value.formatResultType ?? 'none']}
+          TextFiledProps={{
+            select: true,
+            hiddenLabel: true,
+            SelectProps: {
+              autoWidth: true,
+            },
+            value: value.formatResultType || 'none',
+            onChange: (e) => (value.formatResultType = e.target.value as any),
+            children: [
+              <MenuItem key="none" value="none">
+                {t('stayAsIs')}
+              </MenuItem>,
+              <MenuItem key="asHistory" value="asHistory">
+                {t('asHistory')}
+              </MenuItem>,
+            ],
+          }}
+        />
 
         <Box flex={1} />
 
         <Typography variant="body1" component="label">
           {t('variable')}
         </Typography>
-        <TextField
-          hiddenLabel
-          value={value.variable || ''}
-          InputProps={{
-            readOnly,
-            sx: {
-              backgroundColor: { ...getDiffBackground('prepareExecutes', `${value.id}.data.variable`) },
+
+        <IndicatorTextField
+          projectId={projectId}
+          gitRef={gitRef}
+          path={[value.id, value.variable ?? '']}
+          TextFiledProps={{
+            hiddenLabel: true,
+            InputProps: {
+              readOnly,
+              sx: {
+                backgroundColor: { ...getDiffBackground('prepareExecutes', `${value.id}.data.variable`) },
+              },
             },
+            value: value.variable ?? '',
+            onChange: (e) => (value.variable = e.target.value),
           }}
-          onChange={(e) => (value.variable = e.target.value)}
         />
       </Stack>
 
@@ -305,7 +342,7 @@ const ToolDialog = forwardRef<
               const file = store.files[field.value];
               const target = file && isAssistant(file) ? file : undefined;
               const value = target ? { id: target.id, type: target.type, name: target.name } : undefined;
-
+              /* TODO: indicator */
               return (
                 <Autocomplete
                   key={Boolean(field.value).toString()}
@@ -406,6 +443,9 @@ const ToolDialog = forwardRef<
                   render={({ field }) => (
                     <PromptEditorField
                       value={field.value || ''}
+                      projectId={projectId}
+                      gitRef={gitRef}
+                      path={[assistantId, parameter.id]}
                       onChange={(value) => field.onChange({ target: { value } })}
                     />
                   )}